requirements.txt
Flask==2.3.2
google-cloud-storage==2.11.0
google-cloud-pubsub==2.18.0
google-cloud-bigquery==3.11.2
gunicorn==21.0.1  # For production deployment
plotly
